DRAMSimBasedAddress Method - DRAMFaultSimBasedAddress Method

+++ src/memory_system.h
@@ -30,12 +30,12 @@ namespace dramfaultsim {
 
         //data_block_[Channel][Rank][BankGourp][Bank][Row][Col]
         uint64_t ******data_block_;
-        uint8_t recv_addr_channel;
-        uint8_t recv_addr_rank;
-        uint8_t recv_addr_bankgroup;
-        uint8_t recv_addr_bank;
-        uint8_t recv_addr_row;
-        uint8_t recv_addr_column;
+        int recv_addr_channel;
+        int recv_addr_rank;
+        int recv_addr_bankgroup;
+        int recv_addr_bank;
+        int recv_addr_row;
+        int recv_addr_column;
 
         //For FaultModel Return Value
         uint64_t fault_mask;
